Move CPP backend to separate file
